app/recipe/views.py
'''
Views for recipe API
'''
from rest_framework import viewsets
from rest_framework.authentication import TokenAuthentication
from rest_framework.permissions import IsAuthenticated

from core.models import Recipe
from recipe.serializers import RecipeSerializer, RecipeDetailSerializer


class RecipeViewset(viewsets.ModelViewSet):
    '''View for manage recipe APIs.'''
    serializer_class = RecipeDetailSerializer
    queryset = Recipe.objects.all()
    authentication_classes = [TokenAuthentication]
    permission_classes = [IsAuthenticated]

    def get_queryset(self):
        '''Retrive recipes for auth user'''
        return self.queryset.filter(user=self.request.user).order_by('-id')

    def get_serializer_class(self):
        '''Return '''
        if self.action == 'list':
            return RecipeSerializer
        return self.serializer_class    
    
    def perform_create(self, serializer):
        '''create a new recipe'''
        serializer.save(user=self.request.user)


